[ARC] [HACK] Fix ASMCONS pass.
ASMCONS cannot handle CONST (UNSPEC) properly, leading to a wrong
output. Asmcons eventual does reg_overlap_mentioned_p on the input of
the move it created with an input of the asm.  The very first thing
reg_overlap_mentioned_p tests for is if the thing is constant.  And it
is here, and that makes no sense at all. This patch is hacking ASMCONS
and properly handles the CONST (UNSPEC) constructs.

gcc/
2016-12-29  Claudiu Zissulescu  <claziss@synopsys.com>

	* function.c (constant_overlap_mentioned_p): New function.
	(match_asm_constraints_1): Use the above when matching an input
	operand.